% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{check_job_status}
\alias{check_job_status}
\title{Check Status of Submitted Job}
\usage{
check_job_status(job_id)
}
\arguments{
\item{job_id}{Character. SLURM job ID}
}
\value{
Character job status or NA if not found
}
\description{
Check Status of Submitted Job
}
