% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{create_diagnostic_plots}
\alias{create_diagnostic_plots}
\title{Create MCMC Diagnostic Plots}
\usage{
create_diagnostic_plots(..., res = lifecycle::deprecated())
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=plot_mcmc_diagnostics]{plot_mcmc_diagnostics()}}.}

\item{res}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{dpi} instead.}
}
\value{
A \code{ggplot} object of the combined plots, as returned by
\code{\link[=plot_mcmc_diagnostics]{plot_mcmc_diagnostics()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{create_diagnostic_plots()} was renamed to \code{\link[=plot_mcmc_diagnostics]{plot_mcmc_diagnostics()}} so that
every plotting function in the package shares the \verb{plot_*} prefix. Its \code{res}
argument was renamed to \code{dpi} for the same reason.
}
\examples{
chain_files <- c(
  system.file("extdata", "diag_chain1.csv", package = "topolow"),
  system.file("extdata", "diag_chain2.csv", package = "topolow"),
  system.file("extdata", "diag_chain3.csv", package = "topolow")
)
if (all(nzchar(chain_files))) {
  # Deprecated; use plot_mcmc_diagnostics() instead.
  suppressWarnings(
    create_diagnostic_plots(chain_files, mutual_size = 50, save_plot = FALSE)
  )
}

}
\seealso{
\code{\link[=plot_mcmc_diagnostics]{plot_mcmc_diagnostics()}} for the replacement function.
}
\keyword{internal}
