% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{euclidean_dist}
\alias{euclidean_dist}
\title{Calculate Euclidean Distance Between Colors}
\usage{
euclidean_dist(x, y)
}
\arguments{
\item{x, y}{Numeric vectors representing coordinates in LAB color space}
}
\value{
Numeric distance between colors
}
\description{
Calculate Euclidean Distance Between Colors
}
\keyword{internal}
