% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_last_part}
\alias{extract_last_part}
\title{Extract last part of a path or name after final slash}
\usage{
extract_last_part(name)
}
\arguments{
\item{name}{Character string containing path or name}
}
\value{
Last part of string after final slash
}
\description{
Extract last part of a path or name after final slash
}
\keyword{internal}
