% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_year}
\alias{extract_year}
\title{Extract year from sequence name}
\usage{
extract_year(seq_name)
}
\arguments{
\item{seq_name}{Character string containing sequence name}
}
\value{
Extracted year as integer
}
\description{
Extract year from sequence name
}
\keyword{internal}
