% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_sampling.R
\name{find_mode}
\alias{find_mode}
\title{Find Mode of Density Distribution}
\usage{
find_mode(density)
}
\arguments{
\item{density}{List containing density estimate with components:
\describe{
\item{x}{Vector of values}
\item{y}{Vector of density estimates}
}}
}
\value{
Numeric value of the mode
}
\description{
Calculates the mode (maximum point) of a kernel density estimate.
}
