% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_unique_string}
\alias{generate_unique_string}
\title{Generate unique string identifiers with year suffix}
\usage{
generate_unique_string(n, length = 8, lower_bound = 1, upper_bound = 20)
}
\arguments{
\item{n}{Number of strings to generate}

\item{length}{Length of random part of string (default: 8)}

\item{lower_bound}{Lower bound for year suffix (default: 1)}

\item{upper_bound}{Upper bound for year suffix (default: 20)}
}
\value{
Character vector of unique strings with year suffixes
}
\description{
Generate unique string identifiers with year suffix
}
