% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_sampling.R
\name{get_grid}
\alias{get_grid}
\title{Create Grid Around Maximum Likelihood Estimate}
\usage{
get_grid(samples, param, num_points, start_factor, end_factor)
}
\arguments{
\item{samples}{Data frame of MCMC samples with NLL column}

\item{param}{Character name of parameter column}

\item{num_points}{Integer number of points in grid}

\item{start_factor}{Numeric factor for lower bound relative to MLE}

\item{end_factor}{Numeric factor for upper bound relative to MLE}
}
\value{
Numeric vector of grid points
}
\description{
Generates a sequence of values centered on the maximum likelihood estimate (MLE)
of a parameter, extending by specified factors in each direction.
}
\keyword{internal}
