% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_sampling.R
\name{get_sampling_prob_legacy}
\alias{get_sampling_prob_legacy}
\title{Calculate Sampling Distribution Parameters}
\usage{
get_sampling_prob_legacy(samples)
}
\arguments{
\item{samples}{Data frame of previous samples with parameter columns and NLL}
}
\value{
List with:
\item{mean}{Vector of weighted means for each parameter}
\item{cov}{Weighted covariance matrix}
}
\description{
Calculates mean vector and covariance matrix for next sampling distribution
by weighting previous samples according to their likelihoods.
}
\keyword{internal}
