% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggsave}
\alias{ggsave}
\title{Save ggplot with white background}
\usage{
ggsave(..., bg = "white")
}
\arguments{
\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}

\item{bg}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
}
\description{
Wrapper around ggplot2::ggsave that ensures white background.
This function masks ggplot2::ggsave.
}
