% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggsave_white_bg}
\alias{ggsave_white_bg}
\title{Save ggplot with white background}
\usage{
ggsave_white_bg(..., bg = "white")
}
\arguments{
\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}

\item{bg}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
}
\value{
No return value, called for side effects.
}
\description{
Wrapper around \code{ggplot2::ggsave} that ensures a white background by default.
}
