% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h3n2_data}
\alias{h3n2_data}
\title{H3N2 Influenza HI Assay Data from Smith et al. 2004}
\format{
A data frame with the following variables:
\describe{
\item{virusStrain}{Character. Virus strain identifier}
\item{serumStrain}{Character. Antiserum strain identifier}
\item{titer}{Numeric. HI assay titer value}
\item{virusYear}{Numeric. Year virus was isolated}
\item{serumYear}{Numeric. Year serum was collected}
\item{cluster}{Factor. Antigenic cluster assignment}
\item{color}{Character. Color code for visualization}
}
}
\source{
Smith et al. (2004) Science, 305(5682), 371-376.
}
\usage{
h3n2_data
}
\description{
Hemagglutination inhibition (HI) assay data for influenza A/H3N2 viruses
spanning 35 years of evolution.
}
\keyword{datasets}
