% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{has_slurm}
\alias{has_slurm}
\title{Check if SLURM is Available}
\usage{
has_slurm()
}
\value{
Logical indicating if SLURM commands are available
}
\description{
Check if SLURM is Available
}
\keyword{internal}
