% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiv_titers}
\alias{hiv_titers}
\title{HIV Neutralization Assay Data}
\format{
A data frame with the following variables:
\describe{
\item{Antibody}{Character. Antibody identifier}
\item{Virus}{Character. Virus strain identifier}
\item{IC50}{Numeric. IC50 neutralization value}
}
}
\source{
Los Alamos HIV Database (https://www.hiv.lanl.gov/)
}
\usage{
hiv_titers
}
\description{
IC50 neutralization measurements between HIV viruses and antibodies.
}
\keyword{datasets}
