% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiv_viruses}
\alias{hiv_viruses}
\title{HIV Virus Metadata}
\format{
A data frame with the following variables:
\describe{
\item{Virus.name}{Character. Virus strain identifier}
\item{Country}{Character. Country of origin}
\item{Subtype}{Character. HIV subtype}
\item{Year}{Numeric. Year of isolation}
}
}
\source{
Los Alamos HIV Database (https://www.hiv.lanl.gov/)
}
\usage{
hiv_viruses
}
\description{
Reference information for HIV virus strains used in neutralization assays.
}
\keyword{datasets}
