% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{only_virus_vs_as}
\alias{only_virus_vs_as}
\title{Filter matrix to only virus vs antiserum distances}
\usage{
only_virus_vs_as(dist_matrix, selected_names)
}
\arguments{
\item{dist_matrix}{Distance matrix}

\item{selected_names}{Names of selected reference points}
}
\value{
A \code{matrix} of the same dimensions as the input, but with non-virus-vs-antiserum distances set to \code{NA}.
}
\description{
Filter matrix to only virus vs antiserum distances
}
