% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{print.topolow_amcs_diagnostics}
\alias{print.topolow_amcs_diagnostics}
\title{Print Method for Adaptive Monte Carlo Sampling Diagnostics}
\usage{
\method{print}{topolow_amcs_diagnostics}(x, ...)
}
\arguments{
\item{x}{A topolow_amcs_diagnostics object}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for Adaptive Monte Carlo Sampling Diagnostics
}
