% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{print.topolow_diagnostics}
\alias{print.topolow_diagnostics}
\title{Print Method for topolow parameter estimation Diagnostics}
\usage{
\method{print}{topolow_diagnostics}(x, ...)
}
\arguments{
\item{x}{A \code{topolow_diagnostics} object.}

\item{...}{Additional arguments passed to print.}
}
\value{
No return value; called for its side effect of printing a summary.
}
\description{
Print Method for topolow parameter estimation Diagnostics
}
