% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_comparison_helpers.R
\name{run_racmacs}
\alias{run_racmacs}
\title{Run RACMACS Optimization}
\usage{
run_racmacs(truth_matrix, input_matrix, opt_params)
}
\arguments{
\item{truth_matrix}{Matrix of true distances}

\item{input_matrix}{Matrix for optimization}

\item{opt_params}{List of optimization parameters}
}
\value{
List of performance metrics
}
\description{
Runs RACMACS optimization for algorithm comparison.
}
\keyword{internal}
