% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_comparison_helpers.R
\name{run_topolow_local}
\alias{run_topolow_local}
\title{Run Topolow Locally}
\usage{
run_topolow_local(truth_matrix, input_matrix, best_params, opt_params)
}
\arguments{
\item{truth_matrix}{Matrix of true distances}

\item{input_matrix}{Matrix for optimization}

\item{best_params}{List of optimal parameters}

\item{opt_params}{List of optimization parameters}
}
\value{
List of performance metrics
}
\description{
Runs Topolow optimization locally for algorithm comparison.
}
\keyword{internal}
