% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_sampling.R
\name{sample_from_distribution}
\alias{sample_from_distribution}
\title{Sample from Weighted Distribution}
\usage{
sample_from_distribution(dist, n, epsilon)
}
\arguments{
\item{dist}{List with mean and covariance matrix from previous samples}

\item{n}{Integer number of samples to generate}

\item{epsilon}{Numeric probability of sampling from wider distribution}
}
\value{
Data frame of n new samples with columns for each parameter
}
\description{
Generates new samples from a multivariate normal distribution with mean and
covariance determined by weighting previous samples based on their likelihoods.
}
\keyword{internal}
