% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_columns_0_to_10}
\alias{scale_columns_0_to_10}
\title{(Not used anywhere) Scale matrix columns to 0-10 range}
\usage{
scale_columns_0_to_10(mat)
}
\arguments{
\item{mat}{Numeric matrix to scale}
}
\value{
Matrix with columns scaled to 0-10 range
}
\description{
(Not used anywhere) Scale matrix columns to 0-10 range
}
\keyword{internal}
