% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_matrix_by_year}
\alias{sort_matrix_by_year}
\title{Sort symmetric matrix by year}
\usage{
sort_matrix_by_year(matrix)
}
\arguments{
\item{matrix}{symmetric matrix}
}
\value{
Sorted symmetric matrix
}
\description{
Sort symmetric matrix by year
}
\keyword{internal}
