% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{symmetric_to_nonsymmetric_matrix}
\alias{symmetric_to_nonsymmetric_matrix}
\title{Convert distance matrix to assay panel format}
\usage{
symmetric_to_nonsymmetric_matrix(dist_matrix, selected_names)
}
\arguments{
\item{dist_matrix}{Distance matrix}

\item{selected_names}{Names of reference points}
}
\value{
A non-symmetric \code{matrix} in assay panel format, where rows are test antigens and columns are reference antigens.
}
\description{
Convert distance matrix to assay panel format
}
