% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{validate_topolow_df}
\alias{validate_topolow_df}
\title{Validate Input Data Frame}
\usage{
validate_topolow_df(
  df,
  ndim,
  require_clusters = FALSE,
  require_temporal = FALSE
)
}
\arguments{
\item{df}{Data frame to validate}

\item{ndim}{Number of dimensions expected in coordinate columns. Names of coordinate columns must start with a "V".}

\item{require_clusters}{Whether cluster column is required}

\item{require_temporal}{Whether year column is required}
}
\value{
Validated data frame or throws error if invalid
}
\description{
Validates input data frame for visualization functions, checking required
columns and data types.
}
\keyword{internal}
