% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_sampling.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Calculate Weighted Mean}
\usage{
weighted_mean(x, weights)
}
\arguments{
\item{x}{Numeric vector}

\item{weights}{Vector of weights}
}
\value{
Weighted mean value
}
\description{
Helper function to calculate weighted mean of a vector.
}
\keyword{internal}
