% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yy_to_yyyy}
\alias{yy_to_yyyy}
\title{Convert 2-digit to 4-digit year}
\usage{
yy_to_yyyy(dataframe, column_name)
}
\arguments{
\item{dataframe}{Data frame containing year column}

\item{column_name}{Name of year column}
}
\value{
Data frame with converted year column
}
\description{
Convert 2-digit to 4-digit year
}
\keyword{internal}
