% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{color_palettes}
\alias{color_palettes}
\alias{c25}
\title{Color Palettes}
\format{
An object of class \code{character} of length 25.
}
\usage{
c25
}
\description{
Predefined color palettes optimized for visualization.
}
\keyword{datasets}
