% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{coordinates_to_matrix}
\alias{coordinates_to_matrix}
\title{Convert Coordinates to a Distance Matrix}
\usage{
coordinates_to_matrix(positions)
}
\arguments{
\item{positions}{Matrix or Data Frame of coordinates where rows are points and
columns are dimensions.}
}
\value{
A symmetric \code{matrix} of pairwise Euclidean distances between points.
}
\description{
Calculates pairwise Euclidean distances between points in a coordinate space.
}
