% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_sampling.R
\name{generate_kde_samples}
\alias{generate_kde_samples}
\title{Generate New Parameter Samples Using KDE}
\usage{
generate_kde_samples(samples, n, epsilon = 0)
}
\arguments{
\item{samples}{A data frame of previous samples containing parameter columns and a "Holdout_MAE" column.}

\item{n}{The integer number of new samples to generate.}

\item{epsilon}{A numeric probability (0-1) of sampling with a wider bandwidth to
encourage exploration. Default is 0.}
}
\value{
A data frame containing \code{n} new parameter samples.
}
\description{
Generates new parameter samples using weighted kernel density estimation
for each parameter independently. This is an internal helper function for the
adaptive sampling process.
}
\keyword{internal}
