% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{new_layout_config}
\alias{new_layout_config}
\title{Plot Layout Configuration Class}
\usage{
new_layout_config(
  width = 8,
  height = 8,
  dpi = 300,
  aspect_ratio = 1,
  show_grid = TRUE,
  grid_type = "major",
  grid_color = "grey80",
  grid_linetype = "dashed",
  show_axis = TRUE,
  axis_lines = TRUE,
  plot_margin = margin(1, 1, 1, 1, "cm"),
  coord_type = "fixed",
  background_color = "white",
  panel_background_color = "white",
  panel_border = TRUE,
  panel_border_color = "black",
  save_format = "png",
  reverse_x = 1,
  reverse_y = 1,
  x_limits = NULL,
  y_limits = NULL,
  sigma_x = 1,
  sigma_t = 1,
  top_velocity_p = 0.1
)
}
\arguments{
\item{width}{Plot width in inches}

\item{height}{Plot height in inches}

\item{dpi}{Plot resolution}

\item{aspect_ratio}{Plot aspect ratio}

\item{show_grid}{Show plot grid}

\item{grid_type}{Grid type ("none", "major", "minor", "both")}

\item{grid_color}{Grid color}

\item{grid_linetype}{Grid line type}

\item{show_axis}{Show axes}

\item{axis_lines}{Show axis lines}

\item{plot_margin}{Plot margins in cm}

\item{coord_type}{Coordinate type ("fixed", "equal", "flip", "polar")}

\item{background_color}{Plot background color}

\item{panel_background_color}{Panel background color}

\item{panel_border}{Show panel border}

\item{panel_border_color}{Panel border color}

\item{save_format}{Plot save format ("png", "pdf", "svg", "eps")}

\item{reverse_x}{Numeric multiplier for x-axis direction (1 or -1)}

\item{reverse_y}{Numeric multiplier for y-axis direction (1 or -1)}

\item{x_limits}{Numeric vector of length 2 specifying c(min, max) for x-axis. If NULL, limits are set automatically.}

\item{y_limits}{Numeric vector of length 2 specifying c(min, max) for y-axis. If NULL, limits are set automatically.}

\item{sigma_x}{Spatial bandwidth of the kernel function of velocity arrows in plot units}

\item{sigma_t}{Temporal bandwidth of the kernel function of velocity arrows in years}

\item{top_velocity_p}{Top percentile of velocity arrows to draw (between 0 and 1, default: 0.10)}
}
\value{
A layout_config object
}
\description{
S3 class for configuring plot layout including dimensions, margins,
grids and coordinate systems.
}
