% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{print.topolow_convergence}
\alias{print.topolow_convergence}
\title{Print Method for Convergence Diagnostics}
\usage{
\method{print}{topolow_convergence}(x, ...)
}
\arguments{
\item{x}{A topolow_convergence object}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for Convergence Diagnostics
}
