% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_sampling.R
\name{profile_likelihood_result}
\alias{profile_likelihood_result}
\title{S3 Constructor for Profile Likelihood Results}
\usage{
profile_likelihood_result(
  param_values,
  ll_values,
  param_name,
  bandwidth,
  sample_counts
)
}
\arguments{
\item{param_values}{Vector of parameter values.}

\item{ll_values}{Vector of log-likelihood values.}

\item{param_name}{Name of the analyzed parameter.}

\item{bandwidth}{Bandwidth used for local windows.}

\item{sample_counts}{Number of samples per estimate.}
}
\value{
An object of class \code{profile_likelihood}.
}
\description{
Internal S3 constructor for storing results from \code{profile_likelihood}.
}
\keyword{internal}
